#485: refactor RequestObserver to support CoapRequest
